Add JPA CacheLoader and CacheStore implementations
Allows users to use JPA repository beans as Coherence CacheLoader or CacheStore implementations.

- Add example for using Spring JPA repository beans as Coherence cache stores
- Add sample documentation

Two new interfaces have been added: `JpaRepositoryCacheLoader` and `JpaRepositoryCacheStore` that extend JpaRepository.
To create a `CacheLoader` or `CacheStore`, the user just needs to create a new interface that extends one of those interfaces
with the correct generic parameters and annotate it with `@Repository`. Then Spring will then do the rest. Using the Spring
custom namespace in the cache configuration file, this new repository bean can then be injected as the cache store to use for a cache.

A sample has been added in `samples/cachestore-demo`.